backend/requirements.txt
Flask==2.3.2
requests==2.31.0
Flask-Cors==3.0.10
python-dotenv==1.0.0
